Implemented menus, improved hamster rolling
Implemented trinity's new UIs and re-hooked them up to all the code. Cleaned up some of the functionality while I was at it. UIs needed a lot of reorganizing, as a lot of the elements were not zeroed.
Fixed rolling issues that commonly caused odd animations or odd transitions.
Implemented the roll end animation
Updated AR scene with a lot of recent changes, including UI and gamemanager sublevel array